add some documentation and clean up the handler

+++ src/index.ts
@@ -12,38 +12,51 @@ import BigNumber from 'bignumber.js';
 const sqs = new SQS();
 
 interface Registration {
+  // The hash of the label to assign to the address
   labelHash: string;
+  // Address to receive the registrarion
   address: string;
-  value: BigNumber; // Hex value to send with the registration
+  // How much to send
+  value: BigNumber;
+}
+
+interface DeleteMessageInfo {
+  // The URL from which to delete the message
+  QueueUrl: string;
+  // The ID of the message to delete
+  Id: string;
+  // The handle from the receipt of this message
+  ReceiptHandle: string;
 }
 
 export const handler: SQSHandler = async function (event) {
-  const toDelete: { QueueUrl: string; Id: string; ReceiptHandle: string; }[] = [];
-
+  const toDelete: DeleteMessageInfo[] = [];
   const toRegister: Registration[] = [];
 
+  // For each of the records, decide what to do with them.
   await Promise.all(
     event.Records.map(
-      async ({ body, eventSourceARN, messageId, receiptHandle }) => {
+      async message => {
+        const { body, eventSourceARN, messageId, receiptHandle } = message;
+
         try {
           const queueMessage = parseQueueMessage(body);
 
-          let address: string | null = null;
-          try {
-            address = await ethereumProvider.lookupAddress(queueMessage.ensName);
-          } catch (ignored) {
-          }
+          const address = await ethereumProvider.lookupAddress(queueMessage.ensName);
 
           if (address !== null) {
             if (queueMessage.address.toLowerCase() !== address.toLowerCase()) {
-              console.error('message received to register an address that is already registered', queueMessage);
+              console.error(
+                'Message received to register a subdomain that is already registered with a different address',
+                queueMessage
+              );
+            } else {
+              toDelete.push({
+                Id: messageId,
+                ReceiptHandle: receiptHandle,
+                QueueUrl: await getQueueUrl(sqs, eventSourceARN)
+              });
             }
-
-            toDelete.push({
-              Id: messageId,
-              ReceiptHandle: receiptHandle,
-              QueueUrl: await getQueueUrl(sqs, eventSourceARN)
-            });
           } else {
             toRegister.push({
               address: queueMessage.address,
@@ -52,32 +65,38 @@ export const handler: SQSHandler = async function (event) {
             });
           }
         } catch (error) {
-          console.error(error);
-          return null;
+          console.error('Error encountered while processing message', error, message);
         }
       }
     )
   );
 
-  // Group the deletions by the url and send them
-  const grouped = groupBy(toDelete, 'QueueUrl');
+  // Delete any messages from the queue that have already succeeded.
+  if (toDelete.length > 0) {
+    // Group the deletions by the url and send them
+    const grouped = groupBy(toDelete, 'QueueUrl');
 
-  await Promise.all(
-    map(
-      grouped,
-      async (messages, url) =>
-        sqs.deleteMessageBatch({
-          QueueUrl: url,
-          Entries: messages.map(({ Id, ReceiptHandle }) => ({ ReceiptHandle, Id }))
-        })
-    )
-  );
-
-  const gasPrice = await getSafeLowGasPriceWEI();
-
-  const contract = getRegistrarContract();
+    try {
+      await Promise.all(
+        map(
+          grouped,
+          async (messages, url) =>
+            sqs.deleteMessageBatch({
+              QueueUrl: url,
+              Entries: messages.map(({ Id, ReceiptHandle }) => ({ ReceiptHandle, Id }))
+            })
+        )
+      );
+    } catch (error) {
+      console.error('Failed to delete messages from the queue', error, toDelete);
+    }
+  }
 
   if (toRegister.length > 0) {
+    const gasPrice = await getSafeLowGasPriceWEI();
+
+    const contract = getRegistrarContract();
+
     const labels = toRegister.map(({ labelHash }) => labelHash);
     const addresses = toRegister.map(({ address }) => address);
     const values = toRegister.map(({ value }) => `0x${value.toString(16)}`);
@@ -102,6 +121,6 @@ export const handler: SQSHandler = async function (event) {
       console.error('Failed to send transaction', error);
     }
   } else {
-    console.log('No new valid registrations.');
+    console.log('No new registrations. Exiting without submitting transaction.');
   }
 };
